+++ stories/SwitchStory.js
@@ -1,20 +1,20 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
-import Switch from '../src/Switch/Select'
+import Select from '../src/Select/Select'
 
 import ColorModeComparison from './ColorModeComparison'
 
-storiesOf('Switch', module).add('Simple', () => (
+storiesOf('Select', module).add('Simple', () => (
   <div>
     <ColorModeComparison>
-      <Switch options={['First', 'Second']} onSelect={action('Selected')} />
-      <Switch
+      <Select options={['First', 'Second']} onSelect={action('Selected')} />
+      <Select
         options={['First', 'Second']}
         selectedIndex={1}
         onSelect={action('Selected')}
       />
-      <Switch
+      <Select
         options={['First', 'Second', 'Third']}
         onSelect={action('Selected')}
       />
